# openapi.yaml
...
openapi: 3.1.0
info:
  title: 大模型安全评测系统 API
  description: 大模型安全评测系统后端接口骨架。本服务默认监听本机，不包含登录、用户、角色权限和多租户。
  version: 1.0.0
paths:
...
  /api/v1/attack-methods:
    get:
      tags:
      - 攻击方法
      summary: List Attack Methods
      operationId: listAttackMethods
      security:
      - LocalToken: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      - name: enabled
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enabled
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_PaginatedData_AttackMethod__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - 攻击方法
      summary: Create Attack Method
      operationId: createAttackMethod
      security:
      - LocalToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttackMethodCreateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_AttackMethod_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/attack-methods/{method_id}:
    patch:
      tags:
      - 攻击方法
      summary: Update Attack Method
      operationId: updateAttackMethod
      security:
      - LocalToken: []
      parameters:
      - name: method_id
        in: path
        required: true
        schema:
          type: integer
          title: Method Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttackMethodUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_AttackMethod_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - 攻击方法
      summary: Disable Attack Method
      operationId: disableAttackMethod
      security:
      - LocalToken: []
      parameters:
      - name: method_id
        in: path
        required: true
        schema:
          type: integer
          title: Method Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_AttackMethod_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/attack-templates:
    get:
      tags:
      - 攻击模板
      summary: List Attack Templates
      operationId: listAttackTemplates
      security:
      - LocalToken: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      - name: attack_method_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Attack Method Id
      - name: name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_PaginatedData_AttackTemplate__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - 攻击模板
      summary: Create Attack Template
      operationId: createAttackTemplate
      security:
      - LocalToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttackTemplateCreateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_AttackTemplate_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/attack-templates/{template_id}:
    get:
      tags:
      - 攻击模板
      summary: Get Attack Template
      operationId: getAttackTemplate
      security:
      - LocalToken: []
      parameters:
      - name: template_id
        in: path
        required: true
        schema:
          type: integer
          title: Template Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_AttackTemplate_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - 攻击模板
      summary: Update Attack Template
      operationId: updateAttackTemplate
      security:
      - LocalToken: []
      parameters:
      - name: template_id
        in: path
        required: true
        schema:
          type: integer
          title: Template Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttackTemplateUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_AttackTemplate_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - 攻击模板
      summary: Disable Attack Template
      operationId: disableAttackTemplate
      security:
      - LocalToken: []
      parameters:
      - name: template_id
        in: path
        required: true
        schema:
          type: integer
          title: Template Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_AttackTemplate_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/attack-templates/{template_id}/validate:
    post:
      tags:
      - 攻击模板
      summary: Validate Attack Template
      operationId: validateAttackTemplate
      security:
      - LocalToken: []
      parameters:
      - name: template_id
        in: path
        required: true
        schema:
          type: integer
          title: Template Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_AttackTemplateValidationResult_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/attack-templates/{template_id}/preview:
    post:
      tags:
      - 攻击模板
      summary: Preview Attack Template
      operationId: previewAttackTemplate
      security:
      - LocalToken: []
      parameters:
      - name: template_id
        in: path
        required: true
        schema:
          type: integer
          title: Template Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttackTemplatePreviewRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_AttackTemplatePreviewResult_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/attack-templates/{template_id}/versions:
    get:
      tags:
      - 攻击模板
      summary: List Attack Template Versions
      operationId: listAttackTemplateVersions
      security:
      - LocalToken: []
      parameters:
      - name: template_id
        in: path
        required: true
        schema:
          type: integer
          title: Template Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_list_AttackTemplate__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - 攻击模板
      summary: Create Attack Template Version
      operationId: createAttackTemplateVersion
      security:
      - LocalToken: []
      parameters:
      - name: template_id
        in: path
        required: true
        schema:
          type: integer
          title: Template Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttackTemplateVersionCreateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_AttackTemplate_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/attack-templates/{template_id}/publish:
    post:
      tags:
      - 攻击模板
      summary: Publish Attack Template Version
      operationId: publishAttackTemplateVersion
      security:
      - LocalToken: []
      parameters:
      - name: template_id
        in: path
        required: true
        schema:
          type: integer
          title: Template Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_AttackTemplate_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    ApiResponse_AttackMethod_:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/AttackMethod'
          - type: 'null'
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[AttackMethod]
    ApiResponse_AttackTemplatePreviewResult_:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/AttackTemplatePreviewResult'
          - type: 'null'
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[AttackTemplatePreviewResult]
    ApiResponse_AttackTemplateValidationResult_:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/AttackTemplateValidationResult'
          - type: 'null'
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[AttackTemplateValidationResult]
    ApiResponse_AttackTemplate_:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/AttackTemplate'
          - type: 'null'
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[AttackTemplate]
...
    ApiResponse_PaginatedData_AttackMethod__:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/PaginatedData_AttackMethod_'
          - type: 'null'
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[PaginatedData[AttackMethod]]
    ApiResponse_PaginatedData_AttackTemplate__:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/PaginatedData_AttackTemplate_'
          - type: 'null'
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[PaginatedData[AttackTemplate]]
...
    ApiResponse_list_AttackTemplate__:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - items:
              $ref: '#/components/schemas/AttackTemplate'
            type: array
          - type: 'null'
          title: Data
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[list[AttackTemplate]]
...
    AttackMethod:
      properties:
        attack_method_id:
          type: integer
          title: Attack Method Id
        code:
          type: string
          title: Code
        name:
          type: string
          title: Name
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        risk_notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Risk Notes
        enabled:
          type: boolean
          title: Enabled
          default: true
      type: object
      required:
      - attack_method_id
      - code
      - name
      title: AttackMethod
    AttackMethodCreateRequest:
      properties:
        code:
          type: string
          title: Code
        name:
          type: string
          title: Name
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        risk_notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Risk Notes
        enabled:
          type: boolean
          title: Enabled
          default: true
      type: object
      required:
      - code
      - name
      title: AttackMethodCreateRequest
    AttackMethodUpdateRequest:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        risk_notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Risk Notes
        enabled:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enabled
      type: object
      title: AttackMethodUpdateRequest
    AttackTemplate:
      properties:
        attack_template_id:
          type: integer
          title: Attack Template Id
        attack_method_id:
          type: integer
          title: Attack Method Id
        name:
          type: string
          title: Name
        template_text:
          type: string
          title: Template Text
        variables:
          items:
            type: string
          type: array
          title: Variables
        version:
          type: string
          title: Version
          default: v1
        status:
          type: string
          title: Status
          default: active
      type: object
      required:
      - attack_template_id
      - attack_method_id
      - name
      - template_text
      title: AttackTemplate
    AttackTemplateCreateRequest:
      properties:
        attack_method_id:
          type: integer
          title: Attack Method Id
        name:
          type: string
          title: Name
        template_text:
          type: string
          title: Template Text
        variables:
          items:
            type: string
          type: array
          title: Variables
        version:
          type: string
          title: Version
          default: v1
        status:
          type: string
          title: Status
          default: active
      type: object
      required:
      - attack_method_id
      - name
      - template_text
      title: AttackTemplateCreateRequest
    AttackTemplatePreviewRequest:
      properties:
        variables:
          additionalProperties: true
          type: object
          title: Variables
      type: object
      title: AttackTemplatePreviewRequest
    AttackTemplatePreviewResult:
      properties:
        valid:
          type: boolean
          title: Valid
          default: false
        detected_variables:
          items:
            type: string
          type: array
          title: Detected Variables
        missing_variables:
          items:
            type: string
          type: array
          title: Missing Variables
        unused_variables:
          items:
            type: string
          type: array
          title: Unused Variables
        undeclared_variables:
          items:
            type: string
          type: array
          title: Undeclared Variables
        rendered_prompt:
          type: string
          title: Rendered Prompt
          default: ''
      type: object
      title: AttackTemplatePreviewResult
    AttackTemplateUpdateRequest:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        template_text:
          anyOf:
          - type: string
          - type: 'null'
          title: Template Text
        variables:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Variables
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      type: object
      title: AttackTemplateUpdateRequest
    AttackTemplateValidationResult:
      properties:
        valid:
          type: boolean
          title: Valid
          default: false
        detected_variables:
          items:
            type: string
          type: array
          title: Detected Variables
        missing_variables:
          items:
            type: string
          type: array
          title: Missing Variables
        unused_variables:
          items:
            type: string
          type: array
          title: Unused Variables
        undeclared_variables:
          items:
            type: string
          type: array
          title: Undeclared Variables
      type: object
      title: AttackTemplateValidationResult
    AttackTemplateVersionCreateRequest:
      properties:
        version:
          type: string
          title: Version
        template_text:
          anyOf:
          - type: string
          - type: 'null'
          title: Template Text
        variables:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Variables
        status:
          type: string
          title: Status
          default: draft
      type: object
      required:
      - version
      title: AttackTemplateVersionCreateRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PaginatedData_AttackMethod_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/AttackMethod'
          type: array
          title: Items
        page:
          type: integer
          minimum: 1.0
          title: Page
          default: 1
        page_size:
          type: integer
          maximum: 200.0
          minimum: 1.0
          title: Page Size
          default: 20
        total:
          type: integer
          minimum: 0.0
          title: Total
          default: 0
        has_next:
          type: boolean
          title: Has Next
          default: false
      type: object
      title: PaginatedData[AttackMethod]
    PaginatedData_AttackTemplate_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/AttackTemplate'
          type: array
          title: Items
        page:
          type: integer
          minimum: 1.0
          title: Page
          default: 1
        page_size:
          type: integer
          maximum: 200.0
          minimum: 1.0
          title: Page Size
          default: 20
        total:
          type: integer
          minimum: 0.0
          title: Total
          default: 0
        has_next:
          type: boolean
          title: Has Next
          default: false
      type: object
      title: PaginatedData[AttackTemplate]
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    LocalToken:
      type: apiKey
      in: header
      name: X-Local-Token